Refactoring. Added Scheduled task for battery check.

+++ src/main/java/test/drone/dto/CreateDroneDto.java
@@ -9,7 +9,7 @@ import test.drone.entity.Model;
 public record CreateDroneDto(
         @Size(min = 1, max = 100) String serialNumber,
         @NotNull Model model,
-        @Min(0) @Max(500) Integer weightLimit,
+        @Min(0) @Max(500) Double weightLimit,
         @Min(0) @Max(100) Short batteryCapacity
 ) {
 }
