+++ src/main/java/test/drone/dto/DroneDto.java
@@ -12,7 +12,7 @@ import java.util.List;
 public record DroneDto(
         @Size(min = 1, max = 100) String serialNumber,
         @NotNull Model model,
-        @Min(0) @Max(500) Long weightLimit,
+        @Min(0) @Max(500) Double weightLimit,
         @Min(0) @Max(100) Short batteryCapacity,
         @NotNull State state,
         List<MedicationDto> loadedMedications
